organised files and folders removed types files and added types in component file itself added responsiveness for drawer
