--- src/types/Button.ts
@@ -1,9 +1,0 @@
-import { ReactNode } from "react";
-
-export type ButtonProps = {
-    icon?: ReactNode;
-    children: ReactNode;
-    type?: "button" | "submit";
-    color?: "success" | "error";
-    onClick?: () => void;
-};
