--- src/types/Card.ts
@@ -1,4 +1,0 @@
-export type CardProps = {
-    name: string;
-    description: string;
-};
