--- src/types/Drawer.ts
@@ -1,7 +1,0 @@
-import { ReactNode } from "react";
-
-export type DrawerProps = {
-    open: boolean;
-    onClose: () => void;
-    children?: ReactNode;
-};
